refactor: split chain-specific tests into dedicated test files
  Move Ethereum tests to tests/ethereum.rs and Solana tests to
  tests/solana.rs, keeping only generic wallet CRUD tests in wallets.rs.

  GIT_VALID_PII_OVERRIDE
